+++ src/metrics/ml/features/utils.rs
@@ -1,6 +1,6 @@
 use seqair_types::Base;
 
-pub fn one_hot_encode_base(base: impl Into<Base>) -> (f64, f64, f64, f64) {
+pub fn one_hot_encode_base(base: impl Into<Base>) -> (f32, f32, f32, f32) {
     match base.into() {
         Base::A => (1., 0., 0., 0.),
         Base::C => (0., 1., 0., 0.),
@@ -12,6 +12,6 @@ pub fn one_hot_encode_base(base: impl Into<Base>) -> (f64, f64, f64, f64) {
 
 /// Safe division that returns 0.0 instead of NaN when denominator is 0
 #[inline]
-pub fn safe_div(numerator: f64, denominator: f64) -> f64 {
+pub fn safe_div(numerator: f32, denominator: f32) -> f32 {
     if denominator == 0.0 { 0.0 } else { numerator / denominator }
 }
